adding and deleting forms can only have one form up at the same time

+++ XmlDatabase/MainForm.cs
@@ -14,9 +14,10 @@ namespace XmlDatabase
     public partial class MainForm : Form
     {
         private XMLDatabase _xmlDatabase;
+        private bool _addDialogUp = false, _deleteDialogUp = false;
 
-        public XMLDatabase Database => _xmlDatabase;
-
+        public XMLDatabase Database => _xmlDatabase; 
+         
         public string LabelText
         {
             set { if (!string.IsNullOrEmpty(value)) mainLabel.Text = value; }
@@ -63,18 +64,39 @@ namespace XmlDatabase
         {  
             string result = _xmlDatabase.ValidateWithXsd();
             LabelText = result;
-        } 
+        }
+
 
         private void addPersonButton_Click(object sender, EventArgs e)
         {
-            AddPersonForm form = new AddPersonForm(this);
-            form.Show();
+            if (!_addDialogUp)
+            {
+                AddPersonForm form = new AddPersonForm(this);
+                form.Show();
+                form.FormClosed += AddFormClosed;
+                _addDialogUp = true;
+            }
+        }
+
+        private void AddFormClosed(object sender, FormClosedEventArgs e)
+        {
+            _addDialogUp = false;
         }
 
         private void deletePerson_Click(object sender, EventArgs e)
-        { 
-            DeletePersonForm form = new DeletePersonForm(this);
-            form.Show();
+        {
+            if (!_deleteDialogUp)
+            {
+                DeletePersonForm form = new DeletePersonForm(this);
+                form.Show();
+                form.FormClosed += DeleteFormClosed;
+                _deleteDialogUp = true;
+            }
+        }
+
+        private void DeleteFormClosed(object sender, FormClosedEventArgs e)
+        {
+            _deleteDialogUp = false;
         }
 
         private void openXmlButton_Click(object sender, EventArgs e)
@@ -100,6 +122,6 @@ namespace XmlDatabase
                 _xmlDatabase.SetXsdFilePath(dlg.FileName);
                 LabelText = "Załadowano plik XML Schema";
             }
-        }
+        } 
     }
 }
